% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_functions.R
\name{getCoopNet}
\alias{getCoopNet}
\title{getCoopNet}
\usage{
getCoopNet(x)
}
\arguments{
\item{x}{an object of class "PANDA"}
}
\value{
Cooperativity Network
}
\description{
acessor for CoopNet in PANDA object
}
\examples{
data(yeast)
pandaRes <- panda(pandaToyData$motif,
pandaToyData$expression,pandaToyData$ppi,hamming=.5,progress=TRUE)
coopNet <- getCoopNet(pandaRes)
}
