% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_functions.R
\name{getCoRegNet}
\alias{getCoRegNet}
\title{getCoRegNet}
\usage{
getCoRegNet(x)
}
\arguments{
\item{x}{an object of class "PANDA"}
}
\value{
Co-Regulatory Network
}
\description{
acessor for CoRegNet in PANDA object
}
\examples{
data(yeast)
pandaRes <- panda(pandaToyData$motif,
pandaToyData$expression,pandaToyData$ppi,hamming=.5,progress=TRUE)
coregNet <- getCoRegNet(pandaRes)
}
