% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_functions.R
\name{getRegNet}
\alias{getRegNet}
\title{getRegNet}
\usage{
getRegNet(x)
}
\arguments{
\item{x}{an object of class "PANDA"}
}
\value{
Regulatory Network
}
\description{
acessor for RegNet in PANDA object
}
\examples{
data(yeast)
pandaRes <- panda(pandaToyData$motif,
pandaToyData$expression,pandaToyData$ppi,hamming=.5,progress=TRUE)
regNet <- getRegNet(pandaRes)
}
