% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/algorithm_functions.R
\name{panda}
\alias{panda}
\title{Passing Messages between Biological Networks to Refine Predicted Interactions}
\usage{
panda(motif, expr = NULL, ppi = NULL, update.method = "tanimoto",
  alpha = 0.1, hamming = 1e-05, k = NA, output = c("regulatory",
  "coexpression", "cooperative"), z.scale = T, progress = FALSE,
  randomize = "None")
}
\arguments{
\item{motif}{A motif dataset, a data.frame, matrix or exprSet containing 3 columns. Each row describes an motif associated with a transcription factor (column 1) a gene (column 2) and a score (column 3) for the motif.}

\item{expr}{An expression dataset, as a genes (rows) by samples (columns) data.frame}

\item{ppi}{A Protein-Protein interaction dataset, a data.frame containing 3 columns. Each row describes a protein-protein interaction between transcription factor 1(column 1), transcription factor 2 (column 2) and a score (column 3) for the interaction.}

\item{update.method}{function for iterative updating.  This must be one of "tanimoto" and "danmethod"(in development)}

\item{alpha}{value to be used for update variable, alpha (default=0.1)}

\item{hamming}{value at which to terminate the process based on hamming distance (default 10^-5)}

\item{k}{sets the maximum number of iterations PANDA can run before exiting.}

\item{output}{a vector containing which networks to return.  Options include "regulatory", "coregulatory", "cooperative".}

\item{z.scale}{Boolean to indicate use of z-scores in output.  False will use [0,1] scale.}

\item{progress}{Boolean to indicate printing of output for algorithm progress.}

\item{randomize}{method by which to randomize gene expression matrix.  Default "None".  Must be one of "None", "within.gene", "by.genes".  "within.gene" randomization scrambles each row of the gene expression matrix, "by.gene" scrambles gene labels.}
}
\value{
An object of class "panda" containing matrices describing networks achieved by convergence with PANDA algorithm.
"reg.net" is the regulatory network
"coreg.net" is the coregulatory network
"coop.net" is the cooperative network
(Should wrap this up in a special class)
}
\description{
This function runs the PANDA algorithm
}
\examples{
data(yeast)
panda.res.cc <- panda(yeast$motif,yeast$exp.cc,yeast$ppi,hamming=.001,progress=T)
panda.res.sr <- panda(yeast$motif,yeast$exp.sr,yeast$ppi,hamming=.001)
}
\references{
Glass K, Huttenhower C, Quackenbush J, Yuan GC. Passing Messages Between Biological Networks to Refine Predicted Interactions. PLoS One. 2013 May 31;8(5):e64832.
}
\keyword{keywords}

